Allow icon for label
Fixed :focus presentation
Moved mock data to __mock__ folder
