+++ .vscode/settings.json
@@ -30,6 +30,7 @@
     "labeledby",
     "labelledby",
     "nomarker",
+    "nowrap",
     "polarbrightness",
     "popperjs",
     "scrollbar",
