+++ packages/components/src/Form/Inputs/InputFilters/InputFilters.story.tsx
@@ -26,7 +26,7 @@
 
 import React, { useState, FormEvent } from 'react'
 import { Story } from '@storybook/react/types-6-0'
-import { filters } from '../../../__mocks__/sampleInputFilters'
+import { filters } from '../../../__mocks__/filters'
 import { InputText } from '../InputText'
 import { InputFilters, InputFiltersProps, FieldFilter } from './InputFilters'
 import { InputFilterEditorRenderProp } from './inputFilterEditor'
