+++ packages/components/src/Form/Inputs/InputFilters/InputFilters.test.tsx
@@ -27,7 +27,7 @@
 import React, { useState } from 'react'
 import { renderWithTheme } from '@looker/components-test-utils'
 import { fireEvent } from '@testing-library/react'
-import { filters } from '../../../__mocks__/sampleInputFilters'
+import { filters } from '../../../__mocks__/filters'
 import { FieldFilter, InputFilters } from './InputFilters'
 
 const ControlledComponent = () => {
